+++ js-miniapp-sdk/CHANGELOG.md
@@ -1,10 +1,16 @@
 ## CHANGELOG
 
+### 1.7.0 (In PRogress)
+
+- **Feature:** Added support for requesting the display of Banner Ads in Host app. [See here](README.md#Show-Ads).
+
 ### 1.6.1 (2021-01-06)
+
 - **Fix:** Location permission support for iOS/Android Mini App SDK 2.6 and below.
 - **Fix:** Reject promise from `MiniApp.requestLocationPermission` when the user denies location custom permission.
 
 ### 1.6.0 (2020-12-18)
+
 - **Feature:** Added `CustomPermissionName.LOCATION`.
 - **Change:** Updated `requestLocationPermission()` to `requestLocationPermission(permissionDescription?: string)`. From now `requestLocationPermission` will request both custom and device permission respectively. [See here](README.md#Request-Permissions).
 - **Feature:** Added support for requesting Contact list from Host app. [See here](README.md#Requesting-User-details).
@@ -19,7 +25,7 @@
 
 ### 1.3.0 (2020-10-22)
 
-- **Feature:** Added support for requesting the load and display of Interstitial & Rewarded Ads in Host app. [See here](README.md#4-Show-Ads).
+- **Feature:** Added support for requesting the load and display of Interstitial, Rewarded in Host app. [See here](README.md#Show-Ads).
 - **Feature:** Added support for requesting User Name and Profile Photo from Host app. [See here](README.md#Requesting-User-details).
 - **Feature:** Added `MiniApp.getPlatform` for retrieving the platform name of the device. [See here](README.md#check-androidios-device).